+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fix crashing when of union's or interface type's `resolveType` function returns `undefined` or `null` (#731)
 - fix crashing when no reflected type available for fields with params decorators (#724)
 - fix not registering object types implementing interface type when interface type is used as object type field type (#736)
+- properly transform nested array of input type classes (#737)
 
 ## v1.1.0
 ### Features